Lots of stuff from a few weeks worth of daily driving
Over the last few weeks I've brought Emacs back as my daily driver editor
at work. I started this setup quite minimally. Instead of
installing all the bells and whistles out of the gate, I've started
with editor features other than built-in utilities and added plugins as
needed. So far, I've felt the need to add autocompletion for code/text as
well as in menus and a snippet system. Besides that, there are some
general config changes and binding updates.